+++ scripts/config.js
@@ -34,7 +34,7 @@ goog.require('goog.debug.ErrorHandler');
 xhrdav.lib.LIBNAME = 'xhrdavclient';
 
 /** @type {string} */
-xhrdav.lib.VERSION = '0.0.20';
+xhrdav.lib.VERSION = '0.0.21';
 
 /**
  * xhrdavclient Global config