+++ xhrdavclientdeps.js
@@ -4,7 +4,7 @@ goog.addDependency('../../../scripts/config.js', ['xhrdav.lib', 'xhrdav.lib.Conf
 goog.addDependency('../../../scripts/core_ext/string.js', ['xhrdav.lib.string'], ['goog.array', 'goog.functions']);
 goog.addDependency('../../../scripts/core_ext/xmlhttp.js', ['xhrdav.lib.XmlHttp'], ['goog.array']);
 goog.addDependency('../../../scripts/davclient.js', ['xhrdav.lib.Client'], ['goog.Uri', 'goog.dom', 'goog.net.XhrIo', 'xhrdav.lib.Config']);
-goog.addDependency('../../../scripts/davfs.js', ['xhrdav.lib.DavFs'], ['goog.net.XhrManager', 'xhrdav.lib.Client', 'xhrdav.lib.Config', 'xhrdav.lib.ResourceBuilder']);
+goog.addDependency('../../../scripts/davfs.js', ['xhrdav.lib.DavFs'], ['goog.net.XhrManager', 'xhrdav.lib.Client', 'xhrdav.lib.Config', 'xhrdav.lib.ResourceBuilder', 'xhrdav.lib.XhrIoExtBinary']);
 goog.addDependency('../../../scripts/domhandler.js', ['xhrdav.lib.DomHandler'], []);
 goog.addDependency('../../../scripts/domparser.js', ['xhrdav.lib.DomParser'], ['goog.dom.xml', 'goog.functions']);
 goog.addDependency('../../../scripts/errors.js', ['xhrdav.lib.Errors'], ['goog.array', 'goog.json', 'goog.object']);
@@ -15,3 +15,4 @@ goog.addDependency('../../../scripts/httpstatus.js', ['xhrdav.lib.HttpStatus'], 
 goog.addDependency('../../../scripts/resource.js', ['xhrdav.lib.Resource'], ['goog.array', 'goog.object']);
 goog.addDependency('../../../scripts/resourcebuilder.js', ['xhrdav.lib.ResourceBuilder'], ['xhrdav.lib.Config', 'xhrdav.lib.Resource', 'xhrdav.lib.ResourceController']);
 goog.addDependency('../../../scripts/resourcecontroller.js', ['xhrdav.lib.ResourceController'], ['xhrdav.lib.Config', 'xhrdav.lib.Resource']);
+goog.addDependency('../../../scripts/xhrioextbinary.js', ['xhrdav.lib.XhrIoExtBinary'], ['goog.net.XhrIo', 'xhrdav.lib.Config']);